requirements.txt
altgraph==0.17.2
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
click==8.0.1
colorama==0.4.4
colored==1.4.2
dnspython==1.16.0
firebirdsql==1.2.0
Flask==2.0.1
Flask-PyMongo==2.3.0
future==0.18.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
mailjet-rest==1.3.4
MarkupSafe==2.0.1
passlib==1.7.4
pefile==2021.9.3
pycparser==2.20
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
PyJWT==2.1.0
pymongo==3.12.0
python-dotenv==0.19.0
pywin32-ctypes==0.2.0
requests==2.26.0
six==1.16.0
urllib3==1.26.6
validate-docbr==1.8.2
waitress==2.0.0
Werkzeug==2.0.1
wincertstore==0.2
gunicorn==20.0.4
